Add Prime Sieve command-line executable binary

+++ prime-sieve/prime_sieve.cpp
@@ -1,17 +1,19 @@
+#include <cstdint>
+
 class PrimeSieve
 {
 public:
-    PrimeSieve(int size)
+    PrimeSieve(uint64_t size)
     : m_start(2)
     , m_size(size)
     , m_list_primes(nullptr)
     {
         m_not_prime = new bool[size]();
-        for (int i = m_start; i <= size; ++i)
+        for (uint64_t i = m_start; i <= size; ++i)
         {
             if (!m_not_prime[i - m_start])
             {
-                for (int j = m_start; i*j <= size; ++j)
+                for (uint64_t j = m_start; i*j <= size; ++j)
                 {
                     m_not_prime[i*j - m_start] = true;
                 }
@@ -25,7 +27,7 @@ public:
         delete [] m_list_primes;
     }
 
-    bool is_prime(int number)
+    bool is_prime(uint64_t number)
     {
         if (number < m_start)
             return false;
@@ -34,13 +36,13 @@ public:
         return !m_not_prime[number - m_start];
     }
 
-    int* list_primes()
+    uint64_t* list_primes()
     {
         if (!m_list_primes)
         {
-            m_list_primes = new int[m_size]();
-            int ctr = 0;
-            for (int i = 0; i <= m_size; ++i)
+            m_list_primes = new uint64_t[m_size]();
+            uint64_t ctr = 0;
+            for (uint64_t i = 0; i <= m_size; ++i)
             {
                 if (is_prime(i))
                     m_list_primes[ctr++] = i;
@@ -50,9 +52,9 @@ public:
     }
 
 private:
-    const int m_start;
-    const int m_size;
+    const uint64_t m_start;
+    const uint64_t m_size;
     bool* m_not_prime;
-    int* m_list_primes;
+    uint64_t* m_list_primes;
 };
